Split server out from index
